Notification integrated with the screen to change the display time.

+++ app/src/main/java/com/les/povmt/notification/NotificationEventReceiver.java
@@ -23,7 +23,8 @@ public class NotificationEventReceiver extends WakefulBroadcastReceiver {
         AlarmManager alarmManager = (AlarmManager) context.getSystemService(Context.ALARM_SERVICE);
         PendingIntent alarmIntent = getStartPendingIntent(context);
 
-        int time = intent.getIntExtra("time", Integer.MAX_VALUE);
+
+        long time = intent.getLongExtra("time", Integer.MAX_VALUE);
 
         alarmManager.setRepeating(AlarmManager.RTC_WAKEUP,
                 getTriggerAt(new Date()),
